Add files via upload
The new fiber modules are in the attached folder, as well as a folder called “example_fiber” containing the workflow for an experiment of fiber settling in turbulence and the related information in the user manual. Unfortunately, I encountered some issues implementing the fiber orientation module generalized at N>3 cameras, so for now, I left only the fiber tracking option with 2 or 3 cameras.

1) Fiber segmentation
The fibers are now segmented using PCA using the regionprop and labelski functions from the skimage library. This allows a much faster segmentation compared to the previous version.

A new parameter called pca_limit is introduced, which works as a threshold for the image aspect ratio of the segmented fiber. If both particles and fibers are found in an image, a pca_limit > 1 will allow segmenting only fibers. I found it to be useful, especially in some images where both fibers and undesired particles were present.

2) Fiber tracking
The fiber orientation module now works with either 2 or 3 cameras. 

A new parameter called ori_lim is introduced, which gives a maximum error for the orientation reconstruction. This error is used exclusively when more than 2 cameras are used, and it is obtained by computing the angle between a reprojected reconstructed fiber and the segmented fiber in a camera. I put an example in the script on evaluating this threshold error, based on the fiber angular resolution in image space.

3) Fiber smoothing
This module performs smoothing on both positions and orientations and generates a new file containing all the information. 

4) Fiber stitching 
This module performs stitching on fiber centroids, and then reconstruct the orientation with a polynomial fitting. 


A couple of additional notes:
a) I noticed that during the centroid tracking, the code would stop when using the “multiframe” mode when no fibers are matched in a frame (for this reason, I proceeded with the “fourframe” mode instead).
b) I slightly modified the “traj_smoothing_mod.py” file at line 134-135, so that the code automatically trims away the extremes of the smoothed trajectories instead of keeping them.
c) In the workflow, when saving the smoothed and stitched data, I added more decimal values to the saved velocities and accelerations.